final product

+++ src/main/java/com/example/ProductMicroService/dto/ProductDTO.java
@@ -13,8 +13,9 @@ public class ProductDTO {
     Integer categoryId;
     String name;
     String description;
-    Map<String,String> productAttributes;
-    int productRating;
+    Map<String, String> productAttributes;
+    double productRating;
     int noOfSoldUnits;
     String imageUrl;
 }
+
